Add files via upload
Pharmacy Management System application to help pharmacist to manage pharmacy in the systematic ways.
Pharmacy Management System can make the work easier by giving the details of medicine when its name is
entered. A computer gives the details of the medicine like rate of medicine, and expiry date of the medicine. It
becomes very difficult in big medical stores to handle the details of all the medicines manually, so by using this
pharmacy management system We can maintain the records of all the medicines. It is fed with the information
whenever new medicines are brought and it is provided with expire date with search option. When we entire the
name of medicine it gives the details of medicine